ww2global40_balanced2_0 - bomber cost revision
Hi there! this is Regularkid from the AxisandAllies.org forum and co-creator of Balanced Mod. The mod's creators, Adam514 and I, after extensive consultation with the broader Balanced Mod community, have agreed that a revision to bomber costs is necessary to achieve the namesake of the mod--balance.  This change is coded above. Kindly substitute this xml for the existing ww2global40_balanced2_0.xml file. Because github would not allow me to submit a file with a name already in use, I added an "X" to it. Please remove this. If I need to do anything else, please let me know. Thank you!